+++ sdk/messaging/azservicebus/internal/go-amqp/receiver.go
@@ -48,7 +48,7 @@ func (r *Receiver) DrainCredit(ctx context.Context) error {
 // When using ModeSecond, you *must* take an action on the message by calling
 // one of the following: AcceptMessage, RejectMessage, ReleaseMessage, ModifyMessage.
 // When using ModeFirst, the message is spontaneously Accepted at reception.
-func (r *Receiver) Prefetched() (*Message, error) {
+func (r *Receiver) Prefetched() *Message {
 	select {
 	case r.link.ReceiverReady <- struct{}{}:
 	default:
@@ -60,10 +60,10 @@ func (r *Receiver) Prefetched() (*Message, error) {
 	case msg := <-r.link.Messages:
 		log.Debug(3, "Receive() non blocking %d", msg.deliveryID)
 		msg.link = r.link
-		return &msg, nil
+		return &msg
 	default:
 		// done draining messages
-		return nil, nil
+		return nil
 	}
 }
 
@@ -74,10 +74,8 @@ func (r *Receiver) Prefetched() (*Message, error) {
 // one of the following: AcceptMessage, RejectMessage, ReleaseMessage, ModifyMessage.
 // When using ModeFirst, the message is spontaneously Accepted at reception.
 func (r *Receiver) Receive(ctx context.Context) (*Message, error) {
-	msg, err := r.Prefetched()
-
-	if err != nil || msg != nil {
-		return msg, err
+	if msg := r.Prefetched(); msg != nil {
+		return msg, nil
 	}
 
 	// wait for the next message
@@ -281,8 +279,13 @@ func (r *Receiver) sendDisposition(first uint32, last *uint32, state encoding.De
 		State:   state,
 	}
 
-	log.Debug(1, "TX (sendDisposition): %s", fr)
-	return r.link.Session.txFrame(fr, nil)
+	select {
+	case <-r.link.Detached:
+		return r.link.err
+	default:
+		log.Debug(1, "TX (sendDisposition): %s", fr)
+		return r.link.Session.txFrame(fr, nil)
+	}
 }
 
 func (r *Receiver) messageDisposition(ctx context.Context, msg *Message, state encoding.DeliveryState) error {