+++ sdk/messaging/azservicebus/internal/rpc.go
@@ -125,29 +125,29 @@ func NewRPCLink(ctx context.Context, args RPCLinkArgs) (*rpcLink, error) {
 
 	sender, err := session.NewSender(
 		ctx,
-		amqp.LinkTargetAddress(args.Address),
+		args.Address,
+		nil,
 	)
 	if err != nil {
 		return nil, err
 	}
 
-	receiverOpts := []amqp.LinkOption{
-		amqp.LinkSourceAddress(args.Address),
-		amqp.LinkTargetAddress(link.clientAddress),
-		amqp.LinkCredit(defaultReceiverCredits),
+	receiverOpts := &amqp.ReceiverOptions{
+		TargetAddress: link.clientAddress,
+		Credit:        defaultReceiverCredits,
 	}
 
 	if link.sessionID != nil {
 		const name = "com.microsoft:session-filter"
 		const code = uint64(0x00000137000000C)
 		if link.sessionID == nil {
-			receiverOpts = append(receiverOpts, amqp.LinkSourceFilter(name, code, nil))
+			receiverOpts.Filters = append(receiverOpts.Filters, amqp.LinkFilterSource(name, code, nil))
 		} else {
-			receiverOpts = append(receiverOpts, amqp.LinkSourceFilter(name, code, link.sessionID))
+			receiverOpts.Filters = append(receiverOpts.Filters, amqp.LinkFilterSource(name, code, link.sessionID))
 		}
 	}
 
-	receiver, err := session.NewReceiver(ctx, receiverOpts...)
+	receiver, err := session.NewReceiver(ctx, args.Address, receiverOpts)
 	if err != nil {
 		// make sure we close the sender
 		clsCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)